+++ servicio-ventas/src/main/java/com/taller/entity/DetalleVenta.java
@@ -2,6 +2,7 @@ package com.taller.entity;
 
 import jakarta.persistence.*;
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Entity
 @Data
@@ -12,19 +13,17 @@ public class DetalleVenta {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    // FK a la entidad Venta (Tabla D en el mismo servicio)
     @Column(name = "id_venta")
     private Long idVenta;
 
-    // FK a la entidad Producto (Tabla A en otro servicio)
     @Column(name = "id_producto")
     private Long idProducto;
 
     private Integer cantidad;
 
     @Column(precision = 10, scale = 2)
-    private Double precioUnitario; // Precio del producto al momento de la venta
+    private BigDecimal precioUnitario; // Cambiado a BigDecimal
 
     @Column(precision = 10, scale = 2)
-    private Double subtotal;
+    private BigDecimal subtotal; // Cambiado a BigDecimal
 }
